+++ src/Careerbuilder.php
@@ -147,16 +147,6 @@ class Careerbuilder extends AbstractProvider
     }
 
     /**
-     * Get parameters
-     *
-     * @return array
-     */
-    public function getParameters()
-    {
-        return [];
-    }
-
-    /**
      * Get query string for client based on properties
      *
      * @return string