+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All Notable changes to `jobs-careerbuilder` will be documented in this file
 
+## 0.1.5 - 2015-08-12
+
+### Added
+- Nothing
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Using jobs-common 1.0.3
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## 0.1.4 - 2015-07-25
 
 ### Added